MM-8589: expunge global mm_(config|license) from integration components
This also reorganizes `components/integrations/components` as just
`components/integrations`
