+++ components/backstage/backstage_controller.jsx
@@ -7,20 +7,20 @@ import {Route, Switch} from 'react-router-dom';
 
 import Pluggable from 'plugins/pluggable';
 import AnnouncementBar from 'components/announcement_bar';
-import Integrations from 'components/integrations/components/integrations.jsx';
+import Integrations from 'components/integrations';
 import Emoji from 'components/emoji';
 import AddEmoji from 'components/emoji/add_emoji';
-import InstalledIncomingWebhooks from 'components/integrations/components/installed_incoming_webhooks';
-import AddIncomingWehook from 'components/integrations/components/add_incoming_webhook';
-import EditIncomingWebhook from 'components/integrations/components/edit_incoming_webhook';
-import InstalledOutgoingWebhooks from 'components/integrations/components/installed_outgoing_webhooks';
-import AddOutgoingWebhook from 'components/integrations/components/add_outgoing_webhook';
-import EditOutgoingWebhook from 'components/integrations/components/edit_outgoing_webhook';
-import InstalledOauthApps from 'components/integrations/components/installed_oauth_apps';
-import AddOauthApp from 'components/integrations/components/add_oauth_app';
-import EditOauthApp from 'components/integrations/components/edit_oauth_app';
-import CommandsContainer from 'components/integrations/components/commands_container';
-import ConfirmIntegration from 'components/integrations/components/confirm_integration';
+import InstalledIncomingWebhooks from 'components/integrations/installed_incoming_webhooks';
+import AddIncomingWehook from 'components/integrations/add_incoming_webhook';
+import EditIncomingWebhook from 'components/integrations/edit_incoming_webhook';
+import InstalledOutgoingWebhooks from 'components/integrations/installed_outgoing_webhooks';
+import AddOutgoingWebhook from 'components/integrations/add_outgoing_webhook';
+import EditOutgoingWebhook from 'components/integrations/edit_outgoing_webhook';
+import InstalledOauthApps from 'components/integrations/installed_oauth_apps';
+import AddOauthApp from 'components/integrations/add_oauth_app';
+import EditOauthApp from 'components/integrations/edit_oauth_app';
+import CommandsContainer from 'components/integrations/commands_container';
+import ConfirmIntegration from 'components/integrations/confirm_integration';
 
 import BackstageSidebar from './components/backstage_sidebar.jsx';
 import BackstageNavbar from './components/backstage_navbar';
