+++ components/integrations/add_outgoing_webhook/add_outgoing_webhook.jsx
@@ -5,7 +5,7 @@ import PropTypes from 'prop-types';
 import React from 'react';
 
 import {browserHistory} from 'utils/browser_history';
-import AbstractOutgoingWebhook from 'components/integrations/components/abstract_outgoing_webhook.jsx';
+import AbstractOutgoingWebhook from 'components/integrations/abstract_outgoing_webhook.jsx';
 
 const HEADER = {id: 'integrations.add', defaultMessage: 'Add'};
 const FOOTER = {id: 'add_outgoing_webhook.save', defaultMessage: 'Save'};
