+++ components/integrations/commands_container/index.js
@@ -10,10 +10,14 @@ import {getUsers} from 'mattermost-redux/selectors/entities/users';
 import CommandsContainer from './commands_container.jsx';
 
 function mapStateToProps(state, ownProps) {
+    const config = state.entities.general.config;
+    const enableCommands = config.EnableCommands === 'true';
+
     return {
         ...ownProps,
         commands: Object.values(getCommands(state)),
-        users: getUsers(state)
+        users: getUsers(state),
+        enableCommands
     };
 }
