+++ components/integrations/confirm_integration/index.js
