+++ components/integrations/delete_integration.jsx
@@ -5,7 +5,7 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import {FormattedMessage} from 'react-intl';
 
-import DeleteModalTrigger from '../../delete_modal_trigger.jsx';
+import DeleteModalTrigger from 'components/delete_modal_trigger.jsx';
 
 export default class DeleteIntegration extends DeleteModalTrigger {
     get triggerTitle() {
