+++ components/integrations/edit_outgoing_webhook/index.js
@@ -8,13 +8,16 @@ import {getOutgoingHook, updateOutgoingHook} from 'mattermost-redux/actions/inte
 import EditOutgoingWebhook from './edit_outgoing_webhook.jsx';
 
 function mapStateToProps(state, ownProps) {
+    const config = state.entities.general.config;
     const hookId = (new URLSearchParams(ownProps.location.search)).get('id');
+    const enableOutgoingWebhooks = config.EnableOutgoingWebhooks === 'true';
 
     return {
         ...ownProps,
         hookId,
         hook: state.entities.integrations.outgoingHooks[hookId],
-        updateOutgoingHookRequest: state.requests.integrations.createOutgoingHook
+        updateOutgoingHookRequest: state.requests.integrations.createOutgoingHook,
+        enableOutgoingWebhooks
     };
 }
