+++ components/integrations/installed_command.jsx
