+++ components/integrations/installed_commands/index.js
