+++ components/integrations/installed_commands/installed_commands.jsx
