+++ components/integrations/installed_outgoing_webhooks/index.js
@@ -9,18 +9,21 @@ import {getUsers} from 'mattermost-redux/selectors/entities/users';
 import InstalledOutgoingWebhook from './installed_outgoing_webhooks.jsx';
 
 function mapStateToProps(state, ownProps) {
+    const config = state.entities.general.config;
     const teamId = getCurrentTeamId(state);
     const outgoingHooks = getOutgoingHooks(state);
     const outgoingWebhooks = Object.keys(outgoingHooks).
         map((key) => outgoingHooks[key]).
         filter((outgoingWebhook) => outgoingWebhook.team_id === teamId);
+    const enableOutgoingWebhooks = config.EnableOutgoingWebhooks === 'true';
 
     return {
         ...ownProps,
         outgoingWebhooks,
         channels: getAllChannels(state),
         users: getUsers(state),
-        teamId
+        teamId,
+        enableOutgoingWebhooks
     };
 }
