+++ tests/components/integrations/add_incoming_webhook.test.jsx
@@ -4,7 +4,7 @@
 import React from 'react';
 import {shallow} from 'enzyme';
 
-import AddIncomingWebhook from 'components/integrations/components/add_incoming_webhook/add_incoming_webhook.jsx';
+import AddIncomingWebhook from 'components/integrations/add_incoming_webhook/add_incoming_webhook.jsx';
 
 describe('components/integrations/AddIncomingWebhook', () => {
     const createIncomingHook = jest.fn();
